refactor: Update TransactionScreen amount display
This commit updates the `TransactionScreen` to correctly display the transaction amount.

**Key Changes:**

-   **`TransactionScreen.kt`:**
    -   The `Text` composable responsible for displaying the transaction amount now directly uses `transaction.amount` (which is already a Double) instead of `transaction.amount.toDouble()`.

+++ composeApp/src/commonMain/kotlin/dev/balikin/poject/features/transaction/presentation/TransactionScreen.kt
@@ -288,7 +288,7 @@ fun BillCard(
                             )
                         )
                         Text(
-                            text = currencyFormat(transaction.amount.toDouble()),
+                            text = currencyFormat(transaction.amount),
                             style = MaterialTheme.typography.titleMedium.copy(
                                 fontWeight = FontWeight.SemiBold
                             )
